\docType{data}
\name{directionality_reactions}
\alias{directionality_reactions}
\title{List of KEGG reactions with incorrect/inconsistent directionality}
\format{Matrix}
\description{
  This matrix contains a set of KEGG reactions with incorrect/inconsistent directionality.
  The directionality of these reactions has been corrected based on published literature.
  This matrix can be updated or edited by the user if required.
}
\usage{directionality_reactions}
\value{Matrix}
