\docType{data}
\name{hpaNormalTissue}
\alias{hpaNormalTissue}
\title{Expression profiles for proteins in human tissues}
\format{Data.frame}
\description{
  This data frame contains tissue expression data of human proteins, based on the
  Human Protein Atlas project. This data frame was obtained from the hpar package,
  and it is used in MetaboSignal to filter signaling genes based on tissue expression.
}
\usage{data(hpaNormalTissue)}
\value{Data.frame}
