\docType{data}
\name{kegg_pathways}
\alias{kegg_pathways}
\title{Examples of metabolic and signaling human KEGG pathways}
\format{List}
\description{
  This list contains examples of metabolic (first element of the list) and signaling
  (second element of the list) human KEGG pathways.
}
\usage{kegg_pathways}
\value{List}
