\docType{data}
\name{MetaboSignal_table}
\alias{MetaboSignal_table}
\title{Example of MetaboSignal network-table filtered by tissue expression}
\format{Matrix}
\description{
  This network-table was generated using two metabo_paths ("rno00010", "rno00562") and
  two signaling_paths ("rno04910", "rno04151"). Signaling genes were filtered by adipose tissue.
  Notice that due to KEGG/hpar udpates, this network might be different to the one generated
  when running the vignette.
}
\usage{data(MetaboSignal_table)}
\value{Matrix}
