\docType{data}
\name{MetaboSignal_tableUnfiltered}
\alias{MetaboSignal_tableUnfiltered}
\title{Example of MetaboSignal network-table unfiltered by tissue expression}
\format{Matrix}
\description{
  This network-table was generated using two metabo_paths ("rno00010", "rno00562")
  and two signaling_paths ("rno04910", "rno04151"). Signaling genes were not filtered
  by tissue expression.
}
\usage{data(MetaboSignal_tableUnfiltered)}
\value{Matrix}
