\docType{data}
\name{rno_pathways}
\alias{rno_pathways}
\title{Examples of metabolic and signaling rat KEGG pathways}
\format{List}
\description{
  This list contains examples of metabolic (first element of the list) and signaling
  (second element of the list) rat KEGG pathways.
}
\usage{rno_pathways}
\value{List}
