\docType{data}
\name{subnetwork}
\alias{subnetwork}
\title{Shortest-path subnetwork}
\format{Matrix}
\description{
  This network-table corresponds to the bw-ranked shortest-path subnetwork reported
  in the example of the vignette. Notice that due to KEGG/hpar updates, this network
  might be different than the one generated when running the vignette.
}
\usage{data(subnetwork)}
\value{Matrix}
