\docType{data}
\name{kegg_pathways}
\alias{kegg_pathways}
\title{Examples of metabolic and signaling human KEGG pathways}
\format{Matrix}
\description{
  This matrix contains examples of metabolic and signaling human KEGG pathways.
  This matrix was generated with the function "MS_getPathIds( )".
}
\usage{kegg_pathways}
\value{Matrix}
