\docType{data}
\name{MetaboSignal_table}
\alias{MetaboSignal_table}
\title{Example of MetaboSignal network-table}
\format{Matrix}
\description{
  This network-table was generated using two metabo_paths ("rno00010", "rno00562") and
  two signaling_paths ("rno04910", "rno04151"). Notice that due to KEGG udpates,
  this network might be different to the one generated when running the vignette.
}
\usage{data(MetaboSignal_table)}
\value{Matrix}
