\docType{data}
\name{neglected_genes}
\alias{neglected_genes}
\title{Genes removed during tissue-expression filtering}
\format{Vector}
\description{
  This vector contains the IDs of the genes that were removed from the MetaboSignal_table
  due to adipose-tissue filtering in the example of the vignette. Notice that due
  to updates in KEGG/hpar, this vector might be different to the one generated when
  running the vignette.
}
\usage{data(neglected_genes)}
\value{Vector}
